+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ## Bug fixes
 
+* Fixed issue when using a subset of wires with `BasisState`.
+  [(#61)](https://github.com/PennyLaneAI/pennylane-cirq/pull/61)
+
 ## Breaking changes
 
 ## Documentation
@@ -17,7 +20,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Vincent Wong
+Theodor Isacsson, Vincent Wong
 
 ---